move FE utilities to the best place
